moved rating logic to new controller
